+++ database/migrations/2026_07_17_114305_add_version_id_to_publications_table.php
@@ -28,8 +28,10 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('marketing_campaign_post_publications', function (Blueprint $table) {
-            $table->dropForeign('mcp_pub_version_id_fk');
-            $table->dropColumn('marketing_campaign_post_version_id');
+            if (Schema::hasColumn('marketing_campaign_post_publications', 'marketing_campaign_post_version_id')) {
+                $table->dropForeign('mcp_pub_version_id_fk');
+                $table->dropColumn('marketing_campaign_post_version_id');
+            }
         });
     }
 };
